+++ packages/common/src/newsletters/api/index.ts
@@ -1,0 +1,3 @@
+export * from "./subscriptions";
+export * from "./users";
+export * from "./types";