+++ packages/common/src/surveys/utils/get-s123-share-url.ts
@@ -1,0 +1,7 @@
+import { IArcGISContext } from "../../ArcGISContext";
+
+export function getS123ShareUrl(id: string, context: IArcGISContext) {
+  return `${context.survey123Url}/share/${id}?portalUrl=${encodeURIComponent(
+    context.portalUrl
+  )}`;
+}